+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - remove `unsafe` from `Event::get()`.
 - impl `Send` and `Sync` for `Event`.
 - exclude `ComPtr` field access from crate-level API.
+- introduce `ResourceAlignment`, change buffer creation API to reflect it.
 
 # 0.2.1
 - Add a trait `factory::HwndProvider`, to bridge a hwnd provider with the swapchain creation API.